Fix dataframe indexing for sknnr estimators
The previous fix for feature name warnings was to prevent fitting
wrapped estimators with dataframe data, but this had a side effect
of breaking any functionality that depended on fitting with
dataframes, like returning dataframe indexes in sknnr estimators.

I removed the array conversion to restore dataframe indexing, and
added a function that that suppresses missing feature name warnings
in wrapped functions. By wrapping the functions applied by
apply_gufunc (i.e. predict and kneighbors), we can suppress the
warning when it arises at compute time.

I also refactored the utils to avoid some circular dependency issues.
