+++ src/sknnr_spatial/utils/image.py
@@ -1,11 +1,18 @@
+from functools import wraps
+from typing import Callable
+
 import numpy as np
 import xarray as xr
-from typing_extensions import Any
+from typing_extensions import Any, Concatenate, ParamSpec, TypeVar
 
 from ..image._base import ImagePreprocessor, ImageType, ImageWrapper
 from ..image.dataarray import DataArrayWrapper
 from ..image.dataset import DatasetWrapper
 from ..image.ndarray import NDArrayWrapper
+from .wrapper import GenericWrapper
+
+RT = TypeVar("RT")
+P = ParamSpec("P")
 
 
 def is_image_type(X: Any) -> bool:
@@ -18,6 +25,21 @@ def is_image_type(X: Any) -> bool:
         return len(X.dims) == 2
 
     return False
+
+
+def image_or_fallback(
+    func: Callable[Concatenate[GenericWrapper, ImageType, P], RT],
+) -> Callable[Concatenate[GenericWrapper, ImageType, P], RT]:
+    """Decorator that calls the wrapped method for non-image X arrays."""
+
+    @wraps(func)
+    def wrapper(self: GenericWrapper, X_image: ImageType, *args, **kwargs):
+        if not is_image_type(X_image):
+            return getattr(self._wrapped, func.__name__)(X_image, *args, **kwargs)
+
+        return func(self, X_image, *args, **kwargs)
+
+    return wrapper
 
 
 def get_image_wrapper(X_image: ImageType) -> type[ImageWrapper]:
